+++ packages/core/core/embed/embed.test.ts
@@ -1,5 +1,8 @@
 import assert from 'node:assert';
-import { MockEmbeddingModelV1 } from '../test/mock-embedding-model-v1';
+import {
+  MockEmbeddingModelV1,
+  mockEmbed,
+} from '../test/mock-embedding-model-v1';
 import { embed } from './embed';
 
 const dummyEmbedding = [0.1, 0.2, 0.3];
@@ -9,17 +12,24 @@ describe('result.embedding', () => {
   it('should generate embedding', async () => {
     const result = await embed({
       model: new MockEmbeddingModelV1({
-        doEmbed: async ({ values }) => {
-          assert.deepStrictEqual(values, [testValue]);
-
-          return {
-            embeddings: [dummyEmbedding],
-          };
-        },
+        doEmbed: mockEmbed([testValue], [dummyEmbedding]),
       }),
       value: testValue,
     });
 
     assert.deepStrictEqual(result.embedding, dummyEmbedding);
+  });
+});
+
+describe('result.value', () => {
+  it('should include value in the result', async () => {
+    const result = await embed({
+      model: new MockEmbeddingModelV1({
+        doEmbed: mockEmbed([testValue], [dummyEmbedding]),
+      }),
+      value: testValue,
+    });
+
+    assert.deepStrictEqual(result.value, testValue);
   });
 });