+++ packages/core/streams/streaming-text-response.ts
@@ -1,4 +1,4 @@
-import type { ServerResponse } from 'node:http';
+import { mergeStreams } from '../core/util/merge-streams';
 import { StreamData } from './stream-data';
 
 /**
@@ -9,7 +9,7 @@ export class StreamingTextResponse extends Response {
     let processedStream = res;
 
     if (data) {
-      processedStream = res.pipeThrough(data.stream);
+      processedStream = mergeStreams(data.stream, res);
     }
 
     super(processedStream as any, {
@@ -21,31 +21,4 @@ export class StreamingTextResponse extends Response {
       },
     });
   }
-}
-
-/**
- * A utility function to stream a ReadableStream to a Node.js response-like object.
- */
-export function streamToResponse(
-  res: ReadableStream,
-  response: ServerResponse,
-  init?: { headers?: Record<string, string>; status?: number },
-) {
-  response.writeHead(init?.status || 200, {
-    'Content-Type': 'text/plain; charset=utf-8',
-    ...init?.headers,
-  });
-
-  const reader = res.getReader();
-  function read() {
-    reader.read().then(({ done, value }: { done: boolean; value?: any }) => {
-      if (done) {
-        response.end();
-        return;
-      }
-      response.write(value);
-      read();
-    });
-  }
-  read();
 }