+++ packages/google/src/google-generative-ai-settings.ts
@@ -1,5 +1,6 @@
 // https://ai.google.dev/models/gemini
 export type GoogleGenerativeAIModelId =
+  | 'models/gemini-1.5-flash-latest'
   | 'models/gemini-1.5-pro-latest'
   | 'models/gemini-pro'
   | 'models/gemini-pro-vision'